Change mouse=a to mouse=r
mouse=a prevents the ability of copying and pasting out of vim with readable characters.
Change mouse=a to mouse=r and that should fix your issue with that.
one thing I am wondering is, are you changing the config file for your vim with the mouse set to mouse=a?

If mouse=r doesn't give you all the copy past options change it to mouse=v
Both mouse=r and mouse=v have the same functions you are needing, but depending on the vimrc you are using one will work better then the other.

From: https://unix.stackexchange.com/questions/139578/copy-paste-for-vim-is-not-working-when-mouse-set-mouse-a-is-on